Fix /proc/sys/kernel/HZ display.
suse-commit: 20ca0b37ee662a6544e17ec320468bbd31950b33

+++ kernel/sysctl.c
@@ -167,7 +167,7 @@ static void register_proc_table(ctl_table *, struct proc_dir_entry *);
 static void unregister_proc_table(ctl_table *, struct proc_dir_entry *);
 #endif
 
-static unsigned long __HZ = HZ;
+static unsigned int __HZ = HZ;
 /* The default sysctl tables: */
 
 static ctl_table root_table[] = {
@@ -663,7 +663,7 @@ static ctl_table kern_table[] = {
 		.ctl_name	= KERN_HZ, 
 		.procname	= "HZ",
 		.data		= &__HZ,
-		.maxlen		= sizeof(long),
+		.maxlen		= sizeof(int),
 		.mode		= 0444,
 		.proc_handler	= &proc_dointvec,
 	},
